refactor : 인터페이스와 구현체 모듈 분리
