+++ domain/build.gradle.kts
@@ -8,12 +8,6 @@ java {
     targetCompatibility = JavaVersion.VERSION_11
 }
 dependencies {
-    implementation("com.squareup.okhttp3:mockwebserver:4.10.0")
-    implementation("com.squareup.okhttp3:okhttp:4.10.0")
-    implementation("org.json:json:20210307")
-    implementation("com.squareup.retrofit2:retrofit:2.9.0")
-    implementation("com.squareup.retrofit2:converter-gson:2.9.0")
-    implementation("com.squareup.okhttp3:logging-interceptor:4.11.0")
     testImplementation("org.junit.jupiter", "junit-jupiter", "5.8.2")
     testImplementation("org.assertj", "assertj-core", "3.22.0")
 }
