+++ settings.gradle.kts
@@ -15,3 +15,4 @@ dependencyResolutionManagement {
 rootProject.name = "android-shopping-cart"
 include(":app")
 include(":domain")
+include(":data")
